+++ CHANGELOG.md
@@ -17,6 +17,23 @@ Changelog formatting (http://semver.org/):
 ### Security (in case of vulnerabilities.)
 -->
 
+## 3.1.0 (2022-06-29)
+
+### Added
+
+- Create plugin settings with protected post ID option. (df4a90d)
+
+### Changed
+
+- Hide publication controls from non-admin users for select pages. (df4a90d)
+- Convert setup class to use asset loader functions. (71be3d3)
+- Use for kebab case for class handles in place of snake case. (71be3d3)
+- Update NPM and Composer dev-dependencies. (96cc837)
+
+### Removed
+
+- Remove plugin status option for tracking version in favor of asset file version. (dc84044)
+
 ## 3.0.0 (2022-06-02)
 
 ### Breaking changes